prim::pyobject_util.h: use int to index color components to avoid compiler warning
git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@801 e9412239-ebb7-49aa-8f4a-f9aa784616d7

+++ src/prim/pyobject_util.h
@@ -641,7 +641,7 @@ struct PyExportTraits<prim::ColorRGBA>
 		}
 		PyObject** objects = PySequence_Fast_ITEMS(tuple.get());
 		prim::ColorRGBA result;
-		for (Py_ssize_t k = 0; k < size; ++k)
+		for (int k = 0; k < static_cast<int>(size); ++k)
 		{
 			if (!impl::decodeObject(objects[k], result[k], k + 1))
 			{
